+++ src/cpx_io/cpx_system/cpx_ap/cpx_ap_module.py
@@ -22,7 +22,8 @@ class CpxApModule:
         self.base = base
         self.position = position
 
-    def _update_information(self, information):
+    def update_information(self, information):
+        """Update the module information"""
         self.information = information
 
     @CpxBase.require_base