+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
-- [Minor] Unittests for modules
+- [Minor] Unittests for cpxe and cpxap modules
 - [Minor] Added pylintrc file to disable f-string highlighting
 - [Minor] CPXAP: Added read function for AP parameters
 - [Minor] CPXAP: Functionality for IO-Link module